edit student

+++ src/components/Pointstudent/PointstudentItem.js
@@ -1,17 +1,27 @@
 import React, { Component } from 'react'
 import { Link } from 'react-router-dom';
 import { MDBModal, MDBModalBody, MDBModalFooter, MDBBtn, MDBModalHeader } from 'mdbreact';
+import { Modal, ModalHeader, ModalFooter } from 'reactstrap';
  
 export default class PointstudentItem extends Component {
-    state = {
-        modal: false
-      }
-      
-      toggle = () => {
+    constructor(props) {
+        super(props);
+        this.state = ({
+            modal: false,
+            showView: false,
+        });
+        this.showModal = this.showModal.bind(this);
+    }
+    toggle = () => {
         this.setState({
           modal: !this.state.modal
         });
-      }
+    }
+    showModal = () => {
+        this.setState({
+            showView: !this.state.showView
+        });
+    }
     onDelete = (id) => {
    
         this.props.onDelete(id);
@@ -31,29 +41,39 @@ export default class PointstudentItem extends Component {
                 <td>{pointstudent.PointCC}</td>
                 <td>{pointstudent.PointT}</td>
                 <td>
-                    <Link
-                        to={`/pointstudent/${pointstudent.Id}/edit`}
-                        className="btn btn-yellow darken-2"
-                    >
-                        Sửa
-                </Link>
-                     <button
-                        type="button"
-                        className="btn btn-danger"
-                        onClick={this.toggle}
-                    >
-                            Xóa
-                            <MDBModal isOpen={this.state.modal} toggle={this.toggle}>
-                                 <MDBModalHeader toggle={this.toggle}></MDBModalHeader>
-                                 <MDBModalBody>
-                                <span style={{ color: 'black' }}> Bạn chắc chắn muốn xóa bảng điểm <b> {pointstudent.Id}</b>  </span>
-                                </MDBModalBody>
-                                <MDBModalFooter>
-                                <MDBBtn color="secondary" onClick={this.toggle}>Close</MDBBtn>
-                                <MDBBtn color="primary" onClick={() => this.onDelete(pointstudent.id)}>Xóa</MDBBtn>
-                                </MDBModalFooter>
-                            </MDBModal>
-                     </button>
+                <button className="btn-default btn Ripple-parent" onClick={this.showModal}> chức năng
+                <Modal isOpen={this.state.showView} toggle={this.showModal} >
+                    <ModalHeader toggle={this.showModal}>
+                        <b style={{ color: 'black' }}> Bạn chọn chức năng với sinh viên {pointstudent.student.Frist_Name + " " + pointstudent.student.Last_Name}</b>
+                    </ModalHeader>
+                    <ModalFooter>
+                        <Link
+                            to={`/pointstudent/${pointstudent.Id}/edit`}
+                            className="btn btn-yellow darken-2"
+                        >
+                            Sửa
+                        </Link>
+                        <button
+                            type="button"
+                            className="btn btn-danger"
+                            onClick={this.toggle}
+                        >
+                                Xóa
+                                <MDBModal isOpen={this.state.modal} toggle={this.toggle}>
+                                    <MDBModalHeader toggle={this.toggle}></MDBModalHeader>
+                                    <MDBModalBody>
+                                    <span style={{ color: 'black' }}> Bạn chắc chắn muốn xóa bảng điểm <b> {pointstudent.Id}</b>  </span>
+                                    </MDBModalBody>
+                                    <MDBModalFooter>
+                                    <MDBBtn color="secondary" onClick={this.toggle}>Close</MDBBtn>
+                                    <MDBBtn color="primary" onClick={() => this.onDelete(pointstudent.id)}>Xóa</MDBBtn>
+                                    </MDBModalFooter>
+                                </MDBModal>
+                        </button>
+                        <button onClick={this.showModal} className="btn btn-secondary" > đóng </button>
+                    </ModalFooter>
+                </Modal>
+                </button>
                 </td>
             </tr>
         );
